# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /charging/check-in:
    post:
      operationId: checkInChargingSession
      tags: [charging]
      summary: Check a device into charging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargingCheckInInput"
      responses:
        "201":
          description: Charging session and receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargingCheckInResult"
        "409":
          description: Capacity reached or slot conflict
  /charging/verify-claim:
    post:
      operationId: verifyChargingClaim
      tags: [charging]
      summary: Verify a secure charging claim before collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimVerificationInput"
      responses:
        "200":
          description: Minimal staff collection verification details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimVerificationResult"
        "409":
          description: Claim already used or device not ready
  /charging/{id}/status:
    patch:
      operationId: updateChargingStatus
      tags: [charging]
      summary: Update a charging session status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargingStatusInput"
      responses:
        "200":
          description: Updated charging session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargingSession"
  /charging/{id}/collect:
    post:
      operationId: collectChargingSession
      tags: [charging]
      summary: Verify and collect a device
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargingCollectInput"
      responses:
        "200":
          description: Collected charging session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargingSession"
...
  /workspace/register:
    post:
      operationId: registerWorkspaceBooking
      tags: [workspace]
      summary: Register a workspace user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceBookingInput"
      responses:
        "201":
          description: Registered workspace booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceRegistrationResult"
        "409":
          description: Workspace capacity reached
  /workspace/{id}/check-in:
    post:
      operationId: checkInWorkspaceBooking
      tags: [workspace]
      summary: Check a workspace user in
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Checked-in booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceBooking"
  /workspace/{id}/check-out:
    post:
      operationId: checkOutWorkspaceBooking
      tags: [workspace]
      summary: Check a workspace user out
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Checked-out booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceBooking"
  /products:
    get:
      operationId: listProducts
      tags: [inventory]
      summary: List inventory products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [inventory]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List recent sales
      responses:
        "200":
          description: Sale list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a multi-item sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResult"
        "409":
          description: Insufficient inventory
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List ledger transactions
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week, month]
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List recent receipts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Receipt list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
...
  /customer/me:
    get:
      operationId: getCustomerMe
      tags: [customer]
      summary: Get the authenticated customer's profile
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    patch:
      operationId: updateCustomerMe
      tags: [customer]
      summary: Update the authenticated customer's profile and independent preferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerProfileInput" }
      responses:
        "200":
          description: Updated customer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customer/me/charging:
    get:
      operationId: getCustomerMyCharging
      tags: [customer]
      summary: Get only the authenticated customer's charging sessions
      parameters:
        - name: view
          in: query
          schema: { type: string, enum: [active, history, latest] }
      responses:
        "200":
          description: Customer charging activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerChargingResponse" }
...
  /customer/me/receipts:
    get:
      operationId: getCustomerMyReceipts
      tags: [customer]
      summary: Get only the authenticated customer's receipts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Customer receipts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Receipt" }
  /customer/me/receipts/{id}:
    get:
      operationId: getCustomerMyReceiptDetail
      tags: [customer, receipts]
      summary: Get one receipt owned by the authenticated customer
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Owned customer receipt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Receipt" }
...
  /customer/me/notifications/{id}/read:
    patch:
      operationId: markCustomerNotificationRead
      tags: [customer, notifications]
      summary: Mark one owned notification as read
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a visible notification as read
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get SANFAANI settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update administrator-authorized settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsInput" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
...
  /staff/invite:
    post:
      operationId: inviteStaff
      tags: [staff]
      summary: Invite a staff member through the server-side Supabase Admin API
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StaffInviteInput" }
      responses:
        "201":
          description: Invited staff application user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppUser" }
  /staff/{id}/role:
    patch:
      operationId: updateStaffRole
      tags: [staff]
      summary: Change a user role and record its audit history
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleUpdateInput" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
                schema: { $ref: "#/components/schemas/AppUser" }
...
  /receipts/{id}:
    get:
      operationId: getReceiptDetail
      tags: [receipts]
      summary: Get one authorized receipt by identifier or receipt number
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Receipt" }
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  schemas:
...
    Customer:
      type: object
      required: [id, name, whatsappOptIn, notificationPreferences, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        accountStatus: { type: string, enum: [active, inactive] }
        whatsappOptIn: { type: boolean }
        notificationPreferences:
          $ref: "#/components/schemas/NotificationPreferences"
        createdAt: { type: string, format: date-time }
    CustomerInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 2 }
        phone: { type: string }
        email: { type: string }
        whatsappOptIn: { type: boolean }
        notificationPreferences:
          $ref: "#/components/schemas/NotificationPreferences"
...
    ChargingCheckInResult:
      type: object
      required: [session, receipt]
      properties:
        session: { $ref: "#/components/schemas/ChargingSession" }
        receipt: { $ref: "#/components/schemas/Receipt" }
...
    ClaimVerificationInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 32, maxLength: 256 }
...